created parent abstract class CITest
